backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.1
click==8.1.7
dataclasses==0.6
deprecation==2.1.0
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
gotrue==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
packaging==23.2
pluggy==1.3.0
postgrest==0.13.0
py==1.11.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
requests==2.25.1
rfc3986==1.5.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.0
supafunc==0.3.0
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
websockets==10.4
Werkzeug==3.0.1
yarl==1.9.2
